+++ vue/package.json
@@ -9,6 +9,7 @@
   },
   "dependencies": {
     "axios": "^0.18.0",
+    "chrono-node": "^1.3.11",
     "lodash.debounce": "^4.0.8",
     "moment": "^2.24.0",
     "v-autocomplete": "^1.8.2",
